Export supports JSON packages.
This feature adds an new export file type, where export can now
generate either a ZIP File or a JSON file.  The ZIP File generation
is unchanged.

The JSON File provides a simple analog to the compendium `package`
-style exports, allow export of both items and folders.

The output format looks like:
```
{
  "package": { },
  "items": [
    { <Single Document 1> },
    { <Single Document 2> },
    { ... },
    { <Single Document N> },
  ],
  "folders": [
    { <Folder Document 1> },
    { <Folder Document 2> },
    { ... },
    { <Folder Document N> },
  ]
}
```

The set of folders placed into the `folders` array is automatically
discovered from the set of selected documents.  In order to maintain
the correct linkage between the folder and the items, the `_id` field
is injected into each output document.